backend/requirements.txt
# Web framework and server
fastapi==0.128.0
uvicorn[standard]==0.34.0

# Validation and settings
pydantic==2.10.5
pydantic-settings==2.7.0

# Environment
python-dotenv==1.0.1

# Database
asyncpg>=0.31.0

# Redis (Railway / standard)
redis>=5.0.0

# SSE
sse-starlette==2.1.3

# Browser automation
playwright==1.50.0
